requirements.txt
# Core
torch
torchvision
numpy
Pillow
opencv-python
scipy
tqdm
requests

# Model loading
huggingface_hub
safetensors

# 3D / PLY
plyfile
trimesh
open3d>=0.18.0

# Config / utils
omegaconf
einops
jaxtyping
typeguard

# COLMAP export (optional)
# pycolmap==3.10.0

# Gaussian Splatting (required for GS output)
# gsplat>=1.5.3 — install manually, platform-specific wheel:
# Linux:  pip install gsplat @ https://github.com/nerfstudio-project/gsplat/releases/download/v1.5.3/gsplat-1.5.3+pt24cu124-cp310-cp310-linux_x86_64.whl
# Windows: pip install gsplat @ https://github.com/nerfstudio-project/gsplat/releases/download/v1.5.3/gsplat-1.5.3+pt24cu124-cp310-cp310-win_amd64.whl

# CPU offloading (optional, for model_cpu_offload)
accelerate

# PyTorch3D (required for WorldStereo point rendering) — install manually:
# pip install "git+https://github.com/facebookresearch/pytorch3d.git@stable"

# MoGe depth estimator (required for WorldStereo) — install manually:
# pip install "git+https://github.com/microsoft/MoGe.git"

# WorldStereo
diffusers>=0.36.0
transformers>=4.40.0
bitsandbytes>=0.43.0
optimum-quanto>=0.2.4
imageio[ffmpeg]
kornia

# Sky segmentation ONNX fallback (optional)
# onnxruntime

# FP8 quantization (optional, requires compatible GPU)
# torchao

# Video input (optional)
# moviepy==1.0.3


# --- HYRadio Requirements ---
# ComfyUI_HYWorld2 v1.0 dependencies
# IMPORTANT: Do NOT pin torch — ComfyUI manages its own torch version
# Pin loosely to avoid conflicts with other node packs

# Core TTS — Bark via transformers (Gemma 4 also lives here)
transformers>=4.40,<6.0

# Audio processing — WAV I/O and DSP
soundfile>=0.12
numpy>=1.24

# Science news fetching (REQUIRED — no fallback, pipeline will hard-fail without it)
feedparser>=6.0

# Tokenizers (required by transformers, listed for clarity)
tokenizers>=0.15
sentencepiece>=0.1.99

bitsandbytes>=0.42.0
